Basic table creation
First rought draft of what an MVC model could look like using SQlite.
There is a readme file in the SQlite folder, take a look there. You need
to do a few things to get SQlite running with Eclipse.
In the best of worlds, you should be able to fetch the project from the
repo straight from the Eclipse splash screen. Not sure if that's going
to work though.