Better Design
Did a lot of refactoring through all classes, take a look at main to see
how everything works. Much streamlined, such wow.
Still some work to be done for "proper" OOP design, and the unit tests
also.
Hopefully with everything as decoupled as humanly possible that should
not be a problem.